refactor (informationmodel-cat): Move responsibility of harvest metadata update to InformationModelFactory
